added new search and filter feature with updated UI
